Refactor methods to reuse DTOs classes
