+++ students/roberto/services/users/src/routes/handlers/dtos.py
@@ -1,12 +1,31 @@
 from pydantic import BaseModel
 from typing import Dict, List, Optional, AnyStr
 
+# Requests DTOs
+
 
 class CreateUserRequest(BaseModel):
   email: AnyStr
   password: AnyStr
   username: AnyStr
   app_version: Optional[AnyStr] = None
+
+
+class UpdateUserRequest(BaseModel):
+  email: AnyStr
+  password: AnyStr
+  username: AnyStr
+  app_version: Optional[AnyStr] = None
+
+
+class GetUserRequest(BaseModel):
+  id: int
+
+
+class DeleteUserRequest(BaseModel):
+  id: int
+
+# Response DTOs
 
 
 class UserResponse(BaseModel):
@@ -25,3 +44,8 @@ class PaginationMetadata(BaseModel):
 class ListUsersResponse(BaseModel):
   data: List[UserResponse]
   meta: PaginationMetadata
+
+
+class DeleteUserResponse(BaseModel):
+  id: int
+  isDeleted: bool
