+++ students/roberto/services/users/src/routes/handlers/get_user.py
@@ -1,28 +1,16 @@
-from pydantic import BaseModel
-from typing import Optional, AnyStr
+from .dtos import UserResponse, GetUserRequest
 
 
-class GetUserRequest(BaseModel):
-  id: int
-
-
-class GetUserResponse(BaseModel):
-  id: int
-  email: AnyStr
-  username: AnyStr
-  app_version: Optional[AnyStr] = None
-
-
-def get_user(request: GetUserRequest) -> GetUserResponse:
+def get_user(request: GetUserRequest) -> UserResponse:
   userExists = True  # Check if user exists based on id
   if userExists:
-    return GetUserResponse(id=request,
-                           email="Asociated mail",
-                           username="Asociated username",
-                           app_version="v0.0"
-                           )
+    return UserResponse(id=request,
+                        email="Asociated mail",
+                        username="Asociated username",
+                        app_version="v0.0"
+                        )
   else:
-    return GetUserResponse(
+    return UserResponse(
         id=0,
         email="",
         username=""
